change order of grid columns
